+++ src/array/codec/array_to_array.rs
@@ -1,4 +1,6 @@
 //! Array to array codecs.
 
+#[cfg(feature = "bitround")]
+pub mod bitround;
 #[cfg(feature = "transpose")]
 pub mod transpose;